Bug fix for runQuery on reportType `p (Process)
added a flip operation to the reportType~`p` condition in runQuery. lack of flipping caused an error when trying to output the result as a dictionary. Also, temporarily removed the "else throw an error" from the switch in applyPContext, since the error seemed to be getting thrown even when the "else" condition wasn't reached. I'll review it later when I get to validations.