Formatting for p and pm + bugfixes
Fixed some bugs with both process and process model reports. also finished formatting for results of process reports. but in the process of fixing things, I seem to have inadvertently added a bug wherein the results are in the form of a list. that will need to be fixed in a subsequent commit.