Overhaul of Process and Process Model
Based on my session today with @bsulliv , I reworked almost all the code, fixing a bunch of bugs and making the whole thing more readable and maintainable (I hope). I think it should do pretty much everything that my code did at the start of today, but faster and without all of the bugs that I was running into this morning. So, at this point, it should output formatted results for Process Model and Process reports. Validation and Paging Info have not been implemented at this point, though the paging info does show up in the result, since that is part of a properly formatted result according to the spec.